Changed amount in trade to Double, add validation

+++ server/src/main/java/sk/ystad/controllers/PositionController.java
@@ -49,7 +49,7 @@ public class PositionController {
                           @PathVariable(value="symbol") String symbol,
                           @RequestParam("timestamp") String timestamp,
                           @RequestParam("price") Double price,
-                          @RequestParam("amount") int amount,
+                          @RequestParam("amount") double amount,
                           Principal principal){
         return positionService.addTrade(portfolioId, symbol, timestamp, price, amount);
     }
